Fixed bridge with ctx on windows
Fixes: https://github.com/varlink/go/issues/22
